Use `https://geoblacklight.org` as Jekyll baseurl
Jekyll has been generating links throughout the GeoBlacklight site as
`http://geoblacklight.org/`, which is set as the baseurl for the site.
This leads to notifications from Netlify on every deploy that the site
has mixed content, which it doesn't because it has HTTPS enabled on
Netlify.

This pull request modifies a Jekyll configuration option to treat
`https://geoblacklight.org` as the baseurl for everything, which will
suppress these build messages from Netlify.